# optimize with hash160 & generate wif