Fiz algumas alterações que melhoram o desempenho na busca da chave privada.

Quando fazemos a comparação diretamente pelo endereço de uma carteira, acabamos perdendo desempenho no calculo da geração do endereço da carteira. Por isso, minha alteração consiste em ao invés de comparar pelo endereço, comparar pela Hash160 que é um passo anterior a geração do endereço. Economizando tempo e desperdício de processamento.

Para isso, quando as wallets são importadas no programa, fazemos um decode de cada endereço para armazenar o hash160 ao invés do endereço.

Na comparação da private key, geramos até a hash160 para realizar a comparação.

Abaixo deixo uma tabela de comparação de desempenho entre '**Geração e Comparação de Endereço**' e '**Geração e Comparação de Hash160**':

## Comparação de Desempenho
Processador utilizado no teste: Intel i7 6700K
Busca pela chave da carteira 23

### Geração e Comparação de Endereço:
| Keys/s  | Elapsed |
| ------------- | ------------- |
| 233,796  | 6.01s  |
| 234,502  | 5.99s  |
| 234,049  | 6.00s  |
| 229,774  | 6.11s |
| 233,641  | 6.01s  |

Média de 233152 chaves por segundo em 6.02s

### Geração e Comparação de Hash160:
| Keys/s  | Elapsed |
| ------------- | ------------- |
| 295,788  | 4.76s  |
| 291,937  | 4.82s  |
| 290,721  | 4.84s  |
| 297,763  | 4.72s |
| 293,816  | 4.79s  |

Média de 294005 chaves por segundo em 4.78s
Um ganho de aproximadamente 26,10%

----------------
### Geração de WIF
Além disso, também criei um método para geração da WIF quando encontrar a chave privada para uma carteira:
![image](https://github.com/lmajowka/btcgo/assets/817655/622b1158-3fe4-41e8-bbec-2e699544af1d)

E organizei os métodos do base58 em um pacote separado: `/crypto/base58` para ficar menos poluído o pacote principal `main`